% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_per_sample.R
\name{BoxplotsPerSample}
\alias{BoxplotsPerSample}
\title{Draw boxplots with read counts per genes for each sample}
\usage{
BoxplotsPerSample(counts, sample.desc, sample.label.col = NULL,
  xlab = "Raw counts", main = "Box plots per sample: raw counts",
  plot.file = NULL)
}
\arguments{
\item{counts}{a count table with one row per feature and one column per sample}

\item{sample.desc}{sample description table, one row per sample. If it contains a column named "color", this colum will be used to define sample-specific colors.}

\item{sample.label.col=NULL}{column containing sample-associated colors}

\item{xlab="Raw}{counts" X axis label}

\item{main="Box}{plots per sample: raw counts" Main title to display on the plot}

\item{plot.file=NULL}{if specified, the graph will be stored in a pdf-formatted file.}
}
\description{
draw a box plot with one lane per sample (column) of an expression table.
}
\author{
Jacques van Helden
}
