% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_lib.R
\name{CheckRequiredLibraries}
\alias{CheckRequiredLibraries}
\title{check required libraries for DEG analysis, and install them if required}
\usage{
CheckRequiredLibraries(required.libraries, required.bioconductor = NULL)
}
\arguments{
\item{required.libraries}{a vector contianing the names of required CRAN libraries, which will be installed with install.packages()}

\item{required.bioconductor}{a vector containing the required BioConductor libraries, which will be installed with biocLite}
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}
