% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_boxplot.R
\name{count.boxplot}
\alias{count.boxplot}
\title{Draw boxplots with read counts per genes for each sample}
\usage{
count.boxplot(counts, sample.desc, sample.label.col = NULL,
  xlab = "Raw counts", main = "Box plots per sample: raw counts",
  plot.file = NULL)
}
\arguments{
\item{counts}{a count table with one row per feature and one column per sample}

\item{sample.desc}{sample description table, one row per sample}

\item{sample.label.col=NULL}{column containing sample-associated colors}

\item{xlab="Raw}{counts" X axis label}

\item{main="Box}{plots per sample: raw counts" Main title to display on the plot}

\item{plot.file=NULL}{if specified, the graph will be stored in a pdf-formatted file.}
}
\description{
Draw boxplots with read counts per genes for each sample
}
\author{
Jacques van Helden
}
