% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_lib.R
\name{count.correl.heatmap}
\alias{count.correl.heatmap}
\title{Draw a heatmap with the inter-sample correlation matrix.}
\usage{
count.correl.heatmap(count.table, main = NULL, plot.file = NULL,
  score = "cor", cor.method = "pearson", log.transform = TRUE,
  epsilon = 0.1, zlim = NULL, levels = 100, gray.palette = TRUE,
  gamma = 1, plot.values = TRUE, ...)
}
\description{
Draw an equivalent of the t-test volcano plot for RNA-seq
data. The main idea is to display simultaneously the effect size (X axis)
and the significance (Y axis). The measure of the effect size is in this
case the log2 fold change. (logFC).

There is a close relationship between this RNA-seq version of the volcano
and the volcano plots commonly used to show the result of a t-test with microarrays.
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}
