% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_analysis.R
\name{deseq2.analysis}
\alias{deseq2.analysis}
\title{DESeq2 analysis}
\usage{
deseq2.analysis(counts, condition, ref.condition = NULL, comparison.prefix,
  title = comparison.prefix, dir.figures = NULL, ...)
}
\arguments{
\item{counts}{a count table sent to DESeq2. Must contain raw counts (not normalized).}

\item{condition}{a vector with the condition associated to each sample. The length of this vector must equal the number of columns of the count table.}

\item{comparison.prefix}{a string with the prefix for output files}

\item{...}{additional parameters are passed to DESeq2::DESeq() function}

\item{ref.condition=NULL}{reference condition for the differential analysis}

\item{title=comparison.prefix}{main title for the plots}

\item{dir.figures=NULL}{optional directory to save figures}
}
\description{
Detect differentially expressed genes (DEG) using the package DESeq2,
and add a few custom columns (e-value, ...) to the result table.
}
\author{
Jacques van Helden
}
