% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_analysis.R
\name{edger.analysis}
\alias{edger.analysis}
\title{edgeR analysis}
\usage{
edger.analysis(counts, condition, ref.condition = NULL,
  comparison.prefix, norm.method = "TMM", title = paste(sep = "_",
  norm.method, comparison.prefix), dir.figures = NULL, verbose = 0,
  ...)
}
\arguments{
\item{counts}{a count table sent to edgeR Must contain raw counts (not normalized).}

\item{condition}{a vector with the condition associated to each sample. The length of this vector must equal the number of columns of the count table.}

\item{comparison.prefix}{a string with the prefix for output files}

\item{...}{additional parameters are passed to edgeR::exactTest() function}

\item{ref.condition=NULL}{reference condition for the differential analysis}

\item{title=comparison.prefix}{main title for the plots}

\item{dir.figures=NULL}{optional directory to save figures}

\item{norm.method="RLE"}{normalisation method. This parameter strongly affects the results! See edgeR documentation for a list of supported methods}

\item{verbose=0}{level of verbosity}
}
\description{
Detect differentially expressed genes (DEG) using the package edgeR,
and add a few custom columns (e-value, ...) to the result table.
}
\author{
Jacques van Helden
}
