% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot.R
\name{ExportPLot}
\alias{ExportPLot}
\title{Export the current device to one or several files with user-specified format(s).}
\usage{
ExportPLot(file.prefix = "PlotExport", export.formats = "pdf",
  width = dev.size(units = "in")[1], height = dev.size(units =
  "in")[1], horizontal = TRUE, dpi = 72, verbose = 0, ...)
}
\arguments{
\item{verbose}{= 0 level of verbosity}

\item{...}{Additional parameters are passed to the export method}

\item{file.prefix="PlotExport"}{}

\item{export.formats="pdf"}{# supported: postscript, jpg, png, bmp, pdf}

\item{width=11}{Width of output file (in inches)}

\item{height=8}{Height of output file (in inches)}

\item{horizontal=TRUE}{parameter passed to postscript()}

\item{dpi=72}{dots per inch. Automatically converted to width and height in pixels for bitmap formats.}
}
\description{
Export the current device to one or several files with user-specified format(s).
}
\examples{
ExportPLot(file.prefix = "test",
   export.formats = c("postscript","jpg","png","bmp","pdf"))
}
\author{
Jacques van Helden
formats.
}
