% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_colors.R
\name{FeatureColors}
\alias{FeatureColors}
\title{define feature colors for scatter plots as a function of either their score or the local density}
\usage{
FeatureColors(palette.type = "gray", scores = NULL, x = NULL, y = NULL,
  levels = 100, end.colors = c("blue", "red"))
}
\arguments{
\item{type="2col"}{type of palette. Suppored: 2col (2-colors gradient), "gray" (gray scale), "dens" (proportional to local density). 
Palette types 2col and gray resuire to specify a score associated to each feature. 
Density palette requires to specify X and Y coordinates.}

\item{x=NULL}{feature abcsissas for density palettes}

\item{y=NULL}{feature ordinates for density palettes. Must have the same length as x.}

\item{levels=100}{number of palette levels}

\item{end.colors=c('blue', 'red')}{Colors and the two extemes of the palette.}
}
\author{
Jacques van Helden
}
