% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_of_counts.R
\name{HistOfCounts}
\alias{HistOfCounts}
\title{histogram of RNA-seq counts}
\usage{
HistOfCounts(counts, breaks = 100, maxPercentile = 95,
  discardZeroRows = FALSE, discardZeros = FALSE,
  main = "Count histogram", ylab = "Features", xlab = paste(sep = "",
  "Counts (axis truncated to percentile ", maxPercentile, ")"),
  legend.cex = 0.85, verbose = 0, ...)
}
\arguments{
\item{counts}{RNA-seq count table or related dataset.}

\item{verbose}{= 0 level of verbosity}

\item{...}{additional parameters are passed to hist()}

\item{breaks=100}{approximate number of breaks}

\item{maxPercentile=95}{maximum percentile to display on the X axis, in order to avoid the effect of outliers.}

\item{discardZeroRows=FALSE}{if TRUE, discard rows containing only zero values before drawing the histogram.}

\item{discardZeros=FALSE}{if TRUE, discard all zero values before drawing the histogram.}

\item{main="Count}{histogram" main title.}

\item{ylab="Features"}{label for Y axis.}

\item{xlab=paste(sep}{= "", "Counts (axis truncated to percentile ", maxPercentile,")") label for X axis.}

\item{legend.cex=0.85}{font size for the legend}

\item{drawMedian}{= TRUE draw an arrow to mark the median value}

\item{drawMean}{= TRUE draw an arrow to mark the mean value}
}
\value{
a list with the computed parameters + the histogram data
}
\description{
draw an histogram with suitable parameters to RNA-seq count table.
}
\author{
Jacques van Helden
}
