% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_figure.R
\name{index.figure}
\alias{index.figure}
\title{generate a figure chunk of R code to insert in an Rmd report}
\usage{
index.figure(name, file, index.text = NULL, chunk.opt = ", eval=TRUE")
}
\arguments{
\item{name}{name of the chunk}

\item{file}{file continaing the figure}

\item{index.text=NULL}{if spcified, the chunk will be appended to it}

\item{chunk_opt=", }{eval=TRUE" a string specifying one or more options to append to the chunk declaration line. Must start with a comma.}
}
\description{
generate a figure chunk of R code to insert in an Rmd report
}
\author{
Jacques van Helden
}
