% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_deg_table.R
\name{init.deg.table}
\alias{init.deg.table}
\title{Initiate a result table with the input counts and derived statistics.}
\usage{
init.deg.table(count.table, samples1, samples2, gene.info = NULL,
  stats = FALSE)
}
\arguments{
\item{count.table}{table with the counts per reads. This table is used to get feature IDs
(gene IDs, gene names, ...) from row names, and thereby ensure consistency between rows
of the count tables and the summary of differential expression.}

\item{samples1}{vector with  sample IDs of the test condition. Must be a subset of the data table column names.}

\item{samples2}{vector with sample IDs of the control condition. Must be a subset of the data table column names.}

\item{gene.info}{optional table with detailed information about each feature (gene).}

\item{stats=FALSE}{add columns with statistics (mean, quartiles, var, ...) for the whole count table and for the respective sample groups}
}
\value{
a data frame with one row per feature and some descriptive columns + optional table-wise and group-wise statistics.
}
\description{
Initiate a result table with the input counts and derived statistics for the whole data table + for the two sample sets selected for differential analysis.
}
