% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_lib.R
\name{libsize.barplot}
\alias{libsize.barplot}
\title{Draw a barplot with the number of reads per sample}
\usage{
libsize.barplot(stats.per.sample,
  main = "Read library sizes (libsum per sample)", plot.file = NULL, ...)
}
\arguments{
\item{stats.per.sample}{a table with the sample-wise statistics, which ca n be produced by RowStats().}

\item{main}{Main title for the plot}
}
\author{
Jacques van Helden
}
