% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libsize_barplot.R
\name{LibsizeBarplot}
\alias{LibsizeBarplot}
\title{Draw a barplot with the number of reads per sample}
\usage{
LibsizeBarplot(counts, sample.labels = colnames(counts),
  sample.colors = NULL, main = "Library sizes", plot.file = NULL, ...)
}
\arguments{
\item{counts}{a count table with 1 row per feature and 1 column per sample}

\item{main}{Main title for the plot}

\item{sample.labels=colnames(counts)}{sample labels (displayed as y legend)}

\item{sample.colors=NULL}{sample-specific colors (e.g. reflecting conditions)}

\item{plot.file=NULL}{Path of a file to store the figure (pdf format).}
}
\author{
Jacques van Helden
}
