% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rna-seq_dataset.R
\name{LoadRNAseqDataset}
\alias{LoadRNAseqDataset}
\title{Load RNA-seq data, metadata and parameters}
\usage{
LoadRNAseqDataset(countFile, configFile, checkSampleIDs = TRUE,
  reorderSamples = NULL, verbose = 1)
}
\arguments{
\item{countFile}{a tab-separated-value file containing the counts of reads for each feature (row) in each sample (column).}

\item{configFile}{a yaml-formatted file containing the parameters, including the path to metadata files (sample description table, design table).}

\item{checkSampleIDs=TRUE}{check that the sample IDS are equal in the row names of the 
sample description table and the column names (header) of the count table. If "FALSE", row names of 
the sample description fable are taken a sample IDs and imposed as column headers for the count table.}

\item{reorderSamples=NULL}{criterion to ensure consistency of sample order between count table and sample description. 
Supported values:  
\itemize{
\item NULL (no reordering is applied), 
\item counts (reorder columns of the  count table to match the rows sample description table),
\item sample_desc (reorder rows of sample description fable to match columns of the count table).
}}

\item{verbose=1}{level or verbosity}
}
\value{
a list with all the required data for further analysis
\itemize{
\item rawCounts the count table (data.frame) with one row per feature and one column per sample
\item filteredCounts data.frame with the filtered counts. Filtering criteria can specified in the config file.
\item sampleDescriptions a data.frame with one row per sample and one column per attribute.
\item design a data.frame with one row per analysis, indicating the reference and test conditions
\item parameters a list with the parameters read from the yaml-formatted configuration file
}
}
\description{
Load an RNA-seq dataset (i.e. a table with raw counts per feature), 
together with sample descriptions, design for differential expression, and parameters
from  yaml formatted file.
}
\author{
Jacques van Helden
}
