% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_stats.R
\name{MarginStats}
\alias{MarginStats}
\title{compute margin statistics (per row or column) on a data.frame or matrix.}
\usage{
MarginStats(x, margin, verbose = 0, selected.stats = NULL)
}
\arguments{
\item{x}{data.frame or matrix}

\item{margin}{supported values: 1 (row-wise statistics) or 2 (column-wise)}

\item{verbose=0}{level of verbosity}

\item{selected.stats=NULL}{if specified, restrict the computation to a subset of the supported statistics.}
}
\value{
a data.frame with one row per entry (row or column) of the input matrix, and one column per statistics.
#' \describe{
 \item{mean}{arithmetic mean}
 \item{median}{}
 \item{sum}{}
 \item{iqr}{inter-quartile range}
 \item{sd}{estimation of population standard deviation returned by R sd() function. }
 \item{var}{estimation of population variance returned by R var() function. }
 \item{min}{}
 \item{perc01}{first percentile}
 \item{perc05}{5th percentile}
 \item{perc10}{10th percentile}
 \item{Q1}{first quartile, also called lower quartile (LQ) or 25th percentile}
 \item{Q3}{third quartile, also called upper quartile (UQ) or 75th percentile}
 \item{perc90}{90th percentile}
 \item{perc95}{95th percentile}
 \item{perc99}{99th percentile}
 \item{zeros}{number of 0 values}
 \item{na.values}{number of NA values}
 \item{inf.values}{number of infinite values}
 \item{non.null}{number of values that are neither null nor NA nor Inf. }
 \item{non.null}{number of values that are neither null nor NA nor Inf. }
 \item{sdiqr}{standardized IQR (data IQR divided by IQR of the Normal function), which can serve as robust estimator of the standard deviation. }
 \item{max.mean.ratio}{ratio between max and mean values. }
 \item{max.sum.ratio}{ratio between max and sum. }
 \item{mean.median.ratio}{ratio between mean and the median, which can be interpreted as skew indicator. }
 \item{below.mean}{number of values below the mean. }
 \item{fract.below.mean}{fraction of values below the mean. }
 \item{Mcounts}{Million counts = round(digits=2, sum/1e6). Classically used for Next-Generation Sequencing libraries.  }
}
}
\author{
Jacques van Helden
}
