% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_peaks_vs_tfbs.R
\name{MultiPeaksVsTFBS}
\alias{MultiPeaksVsTFBS}
\title{Compare several sets of genomic regions (e.g. Chip_seq peaks) with a set of reference genomic sites (e.g. transcription factor binding sites).}
\usage{
MultiPeaksVsTFBS(peakFiles, TFBSFile, PNGFile,
  main = "Mutual coverage plot", peakSetLabels = NULL,
  verbose = FALSE, drawPlots = FALSE, ...)
}
\arguments{
\item{peakFiles}{a vector contianing a list of files. Each file is supposed to contain a set of genomic regions, in bed format.}

\item{peakSetLabels}{a vector of the same size as peakFiles indicating a label for each peak set.}

\item{...}{Additional parameters are passed to plot()}

\item{sites}{path to a bed-formatted file containing the reference TFBSs. Passed to}

\item{main="Mutual}{coverage plot" main title for the plot.}

\item{verbose=FALSE}{if true, print messages during the execution}

\item{drawPlots=FALSE}{if true, draw diagnostic plots}
}
\value{
A table with one row per peak set, and one column per evaluation parameter (those reported by PeaksVsTFBS).
}
\description{
Evaluate and compare the accuracy of several peaksets by comparing each of 
them with a set of reference TFBSs. The comparison betwee a peak set and the TFBSs is done 
with PeaksVsSTBS, and this function recollects the results and presents them in a synthetic way.
}
\details{
First version: 2017-09
}
\examples{
## Load libraries
library(GenomicRanges)

## TEMPORARY: these two functions were borrowed from the roken package, still to be published
source("~/SnakeChunks/scripts/RSnakeChunks/R/bed_to_granges.R")
source("~/SnakeChunks/scripts/RSnakeChunks/R/bed_import.R")
source("~/SnakeChunks/scripts/RSnakeChunks/R/peaks_vs_tfbs.R")

## Define the root directory for this analysis
mainDir <- "~/RegulonHT_results/GSE93506"
setwd(mainDir)

## File containing the reference TF binding sites
TF <- "OmpR"
bedFile <- paste(sep="",TF, "_unique_sites.bed")
TFBSFile <- file.path(mainDir, "RegulonDB", bedFile)

## Define the directory containing the peaks
peakDir <- file.path(mainDir, "results", "peaks", paste(sep="", TF, "_vs_WT"))

## Identify all the peaks in this directory (recursive search in all subfolders)
peakFiles <- list.files(path=peakDir, pattern=c(".*sickle_bowtie2.*\\\\.bed$"), recursive=TRUE, full.names = TRUE)
message("Peak directory contains ", length(peakFiles), " bed files")

## Compare peaks and TFBSs
multiPeakVsRegDB <- MultiPeaksVsTFBS(
  peakFiles, main=TF,
  peakSetLabels = basename(sub(peakFiles, pattern="\\\\.bed$", replacement = "")), 
  TFBSFile, verbose=TRUE, 
  drawPlots=TRUE)

}
\author{
Claire Rioualen (\email{rioualenclaire@gmail.com}) Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}
