% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot.R
\name{OpenPlotDevice}
\alias{OpenPlotDevice}
\title{open plot device with parameters adapted to the specificities of the different device functions.}
\usage{
OpenPlotDevice(file.prefix, fig.format = "pdf", width = 8,
  height = 8, horizontal = TRUE, dpi = 72, verbose = 0, ...)
}
\arguments{
\item{file.prefix}{file prefix (the extension will be concatenated according to format)}

\item{verbose}{= 0 level of verbosity}

\item{fig.format="pdf"}{file firmat. Supported: postscript, jpg, png, bmp, pdf}

\item{width=8}{Width of output file (in inches).}

\item{height=8}{Height of output file (in inches)}

\item{horizontal=TRUE}{parameter passed to postscript()}

\item{dpi=72}{dots per inch. Automatically converted to width and height in pixels for bitmap formats.}
}
\value{
returns the new device
}
\description{
open plot device with parameters adapted to the specificities of the different device functions.
}
\author{
Jacques van Helden
}
