% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_figure.R
\name{ReportFigure}
\alias{ReportFigure}
\title{generate a figure chunk of R code to insert in an Rmd report}
\usage{
ReportFigure(name, figureFile, reportFile, report.text = NULL,
  out.width = "75\%", chunk.opt = ", eval=TRUE")
}
\arguments{
\item{name}{name of the chunk}

\item{figureFile}{path of the figure file}

\item{reportFile}{path of the report file. The link will be computed relative to this file}

\item{out.width="75\%"}{figure width, relative to page width}

\item{report.text=NULL}{if specified, the chunk will be appended to it}

\item{chunk_opt=", eval=TRUE"}{a string specifying one or more options to append to the chunk declaration line. Must start with a comma.}
}
\description{
generate a figure chunk of R code to insert in an Rmd report.
}
\author{
Jacques van Helden
}
