% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.rnaseq.R
\name{rnaseq.volcanoPlot}
\alias{rnaseq.volcanoPlot}
\title{First version: 2015-09
Last modification: 2015-09}
\usage{
rnaseq.volcanoPlot(deg.table, alpha = 0.05, effect.size.col = "log2FC",
  control.type = "p.value", ...)
}
\arguments{
\item{deg.table}{A table with the results of differential expression. Such tables can be produced by DESeq2, edgeR or other packages.}

\item{control.type}{= "p.value" Control type. Can be a p-value or an equivalent estimation of significance (FDR, e-value, FWER, ...). This value will be converted to y=-log10(x) to obtain the values displayed on the Y axis. The alpha threshold applies on the non-log-converted values.}

\item{...}{}

\item{alpha=0.05}{Threshold to declare features (genes) as positive or negative.}

\item{effect.size.col="log2FC"}{Number or name of the column containing the effect size, whose values will be displayed on the X axis.}
}
\description{
First version: 2015-09
Last modification: 2015-09
}
\examples{

}
