% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna-seq_analysis.R
\name{RNAseqAnalysis}
\alias{RNAseqAnalysis}
\title{Normalisation and differential analysis of an RNA-seq count table}
\usage{
RNAseqAnalysis(count.table, configFile, main.dir = getwd(),
  result.dir = "results", norm.methods = c("DESeq2", "TMM",
  "percentile"), count.prefix = NULL, verbose = 1)
}
\arguments{
\item{count.table}{file containing a table with counts of reads per feature (row) in each sample (column).}

\item{configFile}{a yaml-formatted file defining the mandatory + some optional parameteres}

\item{main.dir=getwd()}{directory from which the script runs. Paths are defined relative to this directory.}

\item{result.dir="results"}{directory where the results will be stored. Should be defined relative to main directory.}

\item{count.prefix=NULL}{basename to export the transformed count tables. If not specified, computed automatically from the name of the count table file.}

\item{norm.methods="TMM"}{vector with one or more normalisation methods. Supported: DESeq2, percentile, quantile, TMM, RLE, upperquartile, none.}

\item{norm.percentile=75}{default percentile to estimate sample-wise scaling factors. By default, set to the upper quartile. Will also be converted as p parameter (between 0 and 1) to tune edgeR upperquartile normalisation.}

\item{verbose=1}{level of verbosity}
}
\description{
Workflow for the analysis of differential expression with RNA-seq transcripttome data.
Includes
\itemize{
\item sample-wiste descriptive statistics (central tendency, dispersion, percentiles, ...)
\item between-sample normalisation with different options
\item detection of differentially expressed genes with DESeq2 and edgeR
\item comparisons between results obtained with alternative methods for between-sample normalisation
\item export of all results in tab-separated value files
\item export of figures in pdf and png formats
\item automatic generation of an HTML report (via Rmd) with links to result files and figures
}
}
\author{
Jacques van Helden
}
