% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_stats.R
\name{RowStats}
\alias{RowStats}
\title{compute row-wise statistics on a data.frame or matrix}
\usage{
RowStats(x, verbose = 0)
}
\arguments{
\item{x}{data.frame or matrix}

\item{verbose=0}{level of verbosity}
}
\value{
a data.frame with one row per column of the input matrix, and one column per statistics
}
\author{
Jacques van Helden
}
