% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_stats.R
\name{statsPerRow}
\alias{statsPerRow}
\title{Descriptive statistics on each row of the input matrix}
\usage{
statsPerRow(x, ...)
}
\arguments{
\item{x}{A matrix or data frame}
}
\value{
A data.frame with one row per row of the input matrix, and one column
per computed statistics.
}
\description{
Compute descriptive parameters (central tendency, dispersion)
for each row of a matrix or data frame.
}
\details{
First version: 2015-03
Replaced by MarginStats: 2018-07-27
}
\examples{
## Load example data set from Den Boer, 2009
library(denboer2009)
data(denboer2009.expr)     ## Load expression table
data(denboer2009.pheno)    ## Load phenotypic data
data(denboer2009.group.labels)    ## Load phenotypic data

stats.per.row <- RowStats(denboer2009.expr, verbose = 1)
head(stats.per.row)
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}
