% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_analysis.R
\name{edger.analysis}
\alias{edger.analysis}
\title{edgeR analysis}
\usage{
edger.analysis(counts, condition, test.condition = 1,
  ref.condition = 2, norm.method = "TMM", thresholds = c(padj = 0.05,
  FC = 2), comparison.prefix, title = paste(sep = "_", norm.method,
  comparison.prefix), dir.figures = NULL, verbose = 0, ...)
}
\arguments{
\item{counts}{a count table sent to edgeR Must contain raw counts (not normalized).}

\item{condition}{a vector with the condition associated to each sample. The length of this vector must equal the number of columns of the count table.}

\item{comparison.prefix}{a string with the prefix for output files}

\item{...}{additional parameters are passed to edgeR::exactTest() function}

\item{test.condition=1}{first condition for edegR::exactTest()}

\item{ref.condition=2}{second condition for edgeR::exactTest(}

\item{title=comparison.prefix}{main title for the plots}

\item{dir.figures=NULL}{optional directory to save figures}

\item{norm.method="TMM"}{normalisation method. This parameter strongly affects the results! See edgeR documentation for a list of supported methods}

\item{thresholds=c(padj=0.05, FC=2)}{thresholds on adjusted p-value (padj) and fold change (FC). 
The result table will include columns indicating which feature pass these thresholds.
See DEGtablePostprocessing() for supported thresholds.}

\item{verbose=0}{level of verbosity}
}
\description{
Detect differentially expressed genes (DEG) using the package edgeR,
and add a few custom columns (e-value, ...) to the result table.
}
\author{
Jacques van Helden
}
