% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_count_table.R
\name{NormalizeCountTable}
\alias{NormalizeCountTable}
\title{scaling RNA-seq count table with a variety of methods (including edgeR and DESeq2 methods)}
\usage{
NormalizeCountTable(counts, class.labels, nozero = TRUE, method = "TMM",
  quantile = 0.75, log2 = FALSE, epsilon = 0.1,
  detailed.sample.stats = FALSE, verbose = 1)
}
\arguments{
\item{counts}{a data frame with counts per feature, with one feature (gene, transcript) per row and one sample per column.}

\item{class.labels}{the class labels associated each sample of the count table. Should be provided in order to adapt it in case samples would be suppressed because they have a scaling factor of 0.}

\item{nozero=TRUE}{If TRUE, zero values are ommitted from the data table before computing the column-wise statistics.}

\item{method="quantile"}{normalization method. Optionally, can be specified as a vector with several methods.

Supported normalization methods: none, sum, mean, median, quantile, TMM, RLE, DESeq2.

The "raw" methods simply returns the raw counts (no normalization).

Sum and mean are really not recommended because very sensitive to outliers.
They are implemented only for the sake of comparison.

Quantile: Preferred quantile = 0.75. Quantiles are more robust to outliers than the mean
or sum. The median is sometimes weak in RNA-seq data, the 75th percentile (quantile 0.75)
seems a good tradeoff between robustness and representativity (taking into account a
representative proportion of the total counts per samples).

Median: actually runs quantile-based scaling factor with quantile=0.5.

TMM: trimmed mean of M-values proposed by Robinson and Oshlack (2010), computed via edgeR::calcNormFactors().

RLE: relative log expression proposed by Anders and Huber (2010), computed via edgeR::calcNormFactors().

DESeq2: compute size factors via DESeq2::estimateSizeFactors()}

\item{quantile=0.75}{quantile used as scaling factor when quantile method is selected.}

\item{log2=FALSE}{Apply log2 transformation after sample size correction}

\item{epsilon=0.1}{value added to all counts before applying the log2 transformation
in order to avoid zero counts.}

\item{detailed.sample.stats=FALSE}{compute detailed sample stats (takes some seconds)}
}
\value{
a list with the following elements.
\itemize{
  \item{counts} The original count table
  \item{class.labels} The original class labels
  \item{parameters}  Normalisation parameters
  \item{norm.result} Normalisation results. If several methods are provided, normcounts is a list with one entry per normalisation method.
}
Each normalisation results contains.
\itemize{
  \item{size.factor} Size factor for each sample.
  \item{scaling.factor} Scaling factor for each sample. Scaling.factor equals 1/size.factor
  \item{counts} Normalized counts
}
}
\description{
normalisation of RNA-seq count table.
More precisely this function runs a sample-wise scaling so that all the samples
have the same value for a user-defined scaling parameter.
By default, we use the quantile 0.75 as scaling factor.
}
\author{
Jacques van Helden and Mustafa AbuElqumsan
}
